refactor: Store 엔티티 이동
